pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[tool.distutils.egg_info]
egg_base = "."

[project]
name = "CardiologyGenAI-AgenticRAG"
version = "0.1.0"
description = "Agent-based RAG framework for cardiology protocols"
requires-python = ">=3.11"
authors = [
    { name = "Christian Faccio", email = "christian.faccio@studenti.units.it" },
    { name = "Gaia Saveri", email = "gaia.saveri@gmail.com" }
]

dependencies = [
    "pydantic==2.11.7",
    "qdrant-client==1.15.1",
    "langchain-core==0.3.75",
    "sentence-transformers==5.1.0",
    "faiss-cpu==1.12.0",
    "fastembed==0.7.2",
    "sqlalchemy==2.0.43",
    "psycopg-binary==3.2.10",
    "asyncpg==0.30.0",
    "langgraph==0.6.7",
    "langgraph-checkpoint-postgres==2.0.23",
    "langchain-ollama==0.3.7",
    "langchain-huggingface==0.3.1",
    "transformers==4.56.0",
    "bitsandbytes==0.42.0",
    "accelerate==1.10.1",
    "langchain-community>=0.3.25",
    "langchain-qdrant>=0.2.0",
    "safetensors>=0.4.3",
    "cardiologygenai-coordo @ git+https://github.com/cardiology-gen-ai/cardiology-gen-ai.git@main"
]

[project.optional-dependencies]
other = [
]
